Changed Principal and MainWindow classes to add the system file function
Added a readonly atribute to save the default path and be changed if is need.
Added a second constructor to change the URL.
Added the cargarArchivo method.
Changed the xaml to a 2x2 grid, and added a pathTextBox and a button to select a file.

Almost everything works fine, the default path is not working if it is relative.